Auction Application Initial Commit
Database created - connection established and tables committed.
Listing front end - code for both front end and back end for the listsing of auctions impelemented
Adding items- form for the front end and bakc end code added to allow logged in users to add new items
extra details - if you click on a exisiting lisiting code for the front end is implemented taht shows the user extra details of it such as the comments bids who posted it ect but i need to implem,ement the back end for bidding and commenting.